split todo util function, components files
