+++ src/App.js
@@ -1,7 +1,4 @@
-import { Routes, Route, redirect } from 'react-router-dom';
 import './App.css';
-import Signup from './pages/Signup.js'
-import Signin from './pages/Signin.js';
